add ex3

+++ Vendedor.java
@@ -2,6 +2,6 @@ public class Vendedor {
 
     public String nome;
     public float salario;
-    public int totalDeVendas;
+    public float totalDeVendas;
     
 }
